+++ changelog.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+#### v3.5.0
+
+* `[terminal]` Using forked [go.linenoise](https://github.com/essentialkaos/go-linenoise) package instead original
+* `[terminal]` Added hints support from new version of `go.linenoise`
+* `[fmtc]` Light colors tag (`-`) support
+* `[usage]` Using dark grey color for option values and example description
+* `[tmp]` Added `DefaultDirPerms` and `DefaultFilePerms` global variables for permissions customization
+* `[tmp]` Improved error handling
+
 #### v3.4.2
 
 * `[strutil]` Fixed bug with overflowing int in `Tail` method